requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
filelock==3.15.4
telethon==1.36.0
requests==2.32.3
Pillow==10.4.0